requirements.txt
python-telegram-bot==13.14
yt-dlp
moviepy
openai-whisper
torch
python-dotenv
